+++ components/shared/AppSidebar/sidebar-menu-item.tsx
@@ -2,30 +2,29 @@ import Link from "next/link";
 import React from "react";
 import { UsersIcon } from "@heroicons/react/24/solid";
 import { ChartBarSquareIcon } from "@heroicons/react/24/outline";
-import clsx from "clsx";
 
 interface MenuItemProps {
   title: string;
   icon?: React.ReactNode;
   type?: "repo" | "list";
-  href: string;
+  url: string;
   isActive?: boolean;
 }
-const SidebarMenuItem = ({ title, icon, type, href, isActive }: MenuItemProps) => {
+const SidebarMenuItem = ({ title, icon, type, url: href }: MenuItemProps) => {
   const getIcon = (type: "repo" | "list") => {
     switch (type) {
       case "list":
-        return <UsersIcon className={clsx("w-5 h-5 text-slate-400", isActive && "text-slate-700")} />;
+        return <UsersIcon className="w-5 h-5 text-slate-400" />;
       case "repo":
-        return <ChartBarSquareIcon className={clsx("w-5 h-5 text-slate-400", isActive && "text-slate-700")} />;
+        return <ChartBarSquareIcon className="w-5 h-5 text-slate-400" />;
     }
   };
 
   return (
-    <li className={clsx("list-none py-2 px-3 rounded-md", isActive ? "bg-slate-100" : "")}>
-      <Link href={href} className=" flex gap-2 items-center">
+    <li className="py-2 px-3">
+      <Link title={title} href={href} className="grid grid-cols-[1.25rem,1fr] gap-2 items-center">
         {type && !icon ? getIcon(type) : icon}
-        <span className={clsx("text-slate-700 font-medium", isActive && "text-slate-800")}>{title}</span>
+        <span className="whitespace-nowrap overflow-hidden overflow-ellipsis">{title}</span>
       </Link>
     </li>
   );